cleaned up app.js:
